+++ packages/sdk/src/publish/MessageFactory.ts
@@ -1,4 +1,4 @@
-import { DEFAULT_PARTITION_COUNT, StreamID, UserID, keyToArrayIndex, toEthereumAddress, utf8ToBinary } from '@streamr/utils'
+import { StreamID, UserID, keyToArrayIndex, toEthereumAddress, utf8ToBinary } from '@streamr/utils'
 import random from 'lodash/random'
 import { Authentication } from '../Authentication'
 import { StreamrClientError } from '../StreamrClientError'
@@ -68,7 +68,7 @@ export class MessageFactory {
             throw new StreamrClientError(`You don't have permission to publish to this stream. Using address: ${publisherId}`, 'MISSING_PERMISSION')
         }
 
-        const partitionCount = (await this.streamRegistry.getStream(this.streamId)).getMetadata().partitions ?? DEFAULT_PARTITION_COUNT
+        const partitionCount = (await this.streamRegistry.getStream(this.streamId)).getPartitionCount()
         let partition
         if (explicitPartition !== undefined) {
             if ((explicitPartition < 0 || explicitPartition >= partitionCount)) {